+++ tests/compile_tests/console_only/image.output
@@ -1,1 +1,1 @@
--13  48  72 
+-11  48  72
