+++ tests/compile_tests/console_only/screenimage.output
@@ -1,1 +1,1 @@
--13 
+-11
