+++ tests/compile_tests/image/test.output
@@ -1,1 +1,1 @@
--13  48  72 
+-11  48  72
