+++ tests/compile_tests/screenimage/test.bas
@@ -7,7 +7,7 @@ $ELSE
 
 ' We can't do _SCREENIMAGE on Linux or Mac OS build agents, but we can still
 ' test that it compiles
-PRINT -13;
+PRINT -11;
 SYSTEM
 
 $END IF
