+++ tests/compile_tests/screenimage/test.output
@@ -1,1 +1,1 @@
--13 
+-11
